Update types for vitest 0.31
As of version 0.31, Vitest no longer exposes most of its types on the
`Vi` namespace, rather it exports them directly. This commit updates our
types to consume and extend the updated type exports.

This commit also no longer augments the global `expect` type, as this is
only desired when the user opts in to importing globals from vitest.
Users will need to follow the normal vitest recommendations to get
global types.